packages/agent-framework-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supermemory-agent-framework"
version = "1.0.0"
description = "Memory tools and middleware for Microsoft Agent Framework with supermemory"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["agent-framework", "supermemory", "ai", "memory", "microsoft"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "agent-framework-core>=1.0.0rc3",
    "supermemory>=3.1.0",
    "typing-extensions>=4.0.0",
]

[dependency-groups]
dev = [
    "black>=24.8.0",
    "flake8>=7.1.2",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.1",
]

[project.urls]
Homepage = "https://supermemory.ai"
Repository = "https://github.com/supermemoryai/supermemory"
Documentation = "https://supermemory.ai/docs"

[tool.hatch.build]
include = ["src/*"]

[tool.hatch.build.targets.wheel]
packages = ["src/supermemory_agent_framework"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
